gateway/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4

# HTTP client for RunPod API and Upstash Redis REST
httpx==0.28.1

# Google Cloud
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.20
